NF: move Test21And26 to compat folder
This is actually only related to compat interface. Plus, it will be necessary in
next commit to access a compat protected method
